source/requirements.txt
  # An example requirements file. If your function has other dependencies,
  # add them below
  functions-framework==3.*
  google-cloud-aiplatform
  google-cloud-storage